requirements.txt
google-auth-oauthlib==0.5.2
google-api-python-client==2.55.0
google-auth==2.9.1
Pillow==9.2.0
